+++ lib/glimmer-dsl-web/samples/hello/hello_input_date_time.rb
@@ -22,21 +22,10 @@
 require 'glimmer-dsl-web'
 
 class TimePresenter
-  attr_accessor :date_time, :month_string, :week_string
+  attr_accessor :date_time, :week_string
   
   def initialize
     @date_time = Time.now
-  end
-  
-  def month_string
-    @date_time&.strftime('%Y-%m')
-  end
-  
-  def month_string=(value)
-    if value.match(/^\d{4}-\d{2}$/)
-      year, month = value.split('-')
-      self.date_time = Time.new(year, month, date_time.day, date_time.hour, date_time.min)
-    end
   end
   
   def week_string
@@ -56,6 +45,8 @@ class TimePresenter
       date_parts = date_time_parts.first.split('-')
       time_parts = date_time_parts.last.split(':')
       self.date_time = Time.new(*date_parts, *time_parts)
+    elsif value.strip.empty?
+      self.date_time = nil
     end
   end
 end
@@ -78,22 +69,22 @@ Document.ready? do
       input(id: 'date-field', type: 'date') {
         value <=> [@time_presenter, :date_time]
       }
-      
+
       label('Time: ', for: 'time-field')
       input(id: 'time-field', type: 'time') {
         value <=> [@time_presenter, :date_time]
       }
-      
+
       label('Month: ', for: 'month-field')
       input(id: 'month-field', type: 'month') {
-        value <=> [@time_presenter, :month_string, computed_by: :date_time]
+        value <=> [@time_presenter, :date_time]
       }
-      
+
       label('Week: ', for: 'week-field')
       input(id: 'week-field', type: 'week', disabled: true) {
-        value <=> [@time_presenter, :week_string, computed_by: :date_time]
+        value <= [@time_presenter, :week_string, computed_by: :date_time]
       }
-      
+
       label('Time String: ', for: 'time-string-field')
       input(id: 'time-string-field', type: 'text') {
         value <=> [@time_presenter, :date_time_string, computed_by: :date_time]